+++ db/migrate/20161025162531_add_message_to_checks.rb
@@ -1,0 +1,5 @@
+class AddMessageToChecks < ActiveRecord::Migration[5.0]
+  def change
+    add_column :checks, :message, :string
+  end
+end